Change CI Dependencies Installation to avoid Poetry
