+++ tests/unit_tests/prompts/test_prompt.py
@@ -145,100 +145,3 @@ def test_partial() -> None:
     assert new_result == "This is a 3 test."
     result = prompt.format(foo="foo")
     assert result == "This is a foo test."
-
-
-def test_prompt_from_jinja2_template() -> None:
-    """Test prompts can be constructed from a jinja2 template."""
-    # Empty input variable.
-    template = """Hello there
-There is no variable here {
-Will it get confused{ }? 
-    """
-    prompt = PromptTemplate.from_template(template, template_format="jinja2")
-    expected_prompt = PromptTemplate(
-        template=template, input_variables=[], template_format="jinja2"
-    )
-    assert prompt == expected_prompt
-
-    # Multiple input variables.
-    template = """\
-Hello world
-
-Your variable: {{ foo }}
-
-{# This will not get rendered #}
-
-{% if bar %}
-You just set bar boolean variable to true
-{% endif %}
-
-{% for i in foo_list %}
-{{ i }}
-{% endfor %}
-"""
-    prompt = PromptTemplate.from_template(template, template_format="jinja2")
-    expected_prompt = PromptTemplate(
-        template=template,
-        input_variables=["bar", "foo", "foo_list"],
-        template_format="jinja2",
-    )
-
-    assert prompt == expected_prompt
-
-    # Multiple input variables with repeats.
-    template = """\
-Hello world
-
-Your variable: {{ foo }}
-
-{# This will not get rendered #}
-
-{% if bar %}
-You just set bar boolean variable to true
-{% endif %}
-
-{% for i in foo_list %}
-{{ i }}
-{% endfor %}
-
-{% if bar %}
-Your variable again: {{ foo }}
-{% endif %}
-"""
-    prompt = PromptTemplate.from_template(template, template_format="jinja2")
-    expected_prompt = PromptTemplate(
-        template=template,
-        input_variables=["bar", "foo", "foo_list"],
-        template_format="jinja2",
-    )
-    assert prompt == expected_prompt
-
-
-def test_prompt_jinja2_missing_input_variables() -> None:
-    """Test error is raised when input variables are not provided."""
-    template = "This is a {{ foo }} test."
-    input_variables: list = []
-    with pytest.raises(ValueError):
-        PromptTemplate(
-            input_variables=input_variables, template=template, template_format="jinja2"
-        )
-
-
-def test_prompt_jinja2_extra_input_variables() -> None:
-    """Test error is raised when there are too many input variables."""
-    template = "This is a {{ foo }} test."
-    input_variables = ["foo", "bar"]
-    with pytest.raises(ValueError):
-        PromptTemplate(
-            input_variables=input_variables, template=template, template_format="jinja2"
-        )
-
-
-def test_prompt_jinja2_wrong_input_variables() -> None:
-    """Test error is raised when name of input variable is wrong."""
-    template = "This is a {{ foo }} test."
-    input_variables = ["bar"]
-    with pytest.raises(ValueError):
-        PromptTemplate(
-            input_variables=input_variables, template=template, template_format="jinja2"
-        )
